Embed Commands
spicing it up

+++ Xilla-Discord-Core-v2/src/main/java/net/xilla/discordcore/command/cmd/EmbedCommand.java
@@ -24,37 +24,42 @@ public class EmbedCommand extends CoreObject {
 
             if(data.getInputType().equals(CoreCommandExecutor.discord_input)) {
 
-                MessageReceivedEvent event = (MessageReceivedEvent)data.get();
-                event.getMessage().delete().queue();
-
-                StringBuilder message = new StringBuilder();
-                for (int i = 0; i < data.getArgs().length; i++) {
-                    String word = data.getArgs()[i];
-                    message.append(word);
-                    if (i != data.getArgs().length - 1)
-                        message.append(" ");
-
-                }
-
-                String title = null;
-                String description = null;
-                if (message.toString().startsWith("\"")) {
-                    String temp = message.toString().substring(1);
-                    int index = temp.indexOf("\"");
-                    title = temp.substring(0, index);
-                    description = message.toString().substring(index + 2);
-                }
-
                 EmbedBuilder embedBuilder = new EmbedBuilder();
-
-                if (title != null) {
-                    embedBuilder.setTitle(title);
-                    embedBuilder.setDescription(description);
-                } else {
-                    embedBuilder.setDescription(message.toString());
-                }
-
                 embedBuilder.setColor(Color.decode(getDiscordCore().getCoreSetting().getEmbedColor()));
+                if(data.getArgs().length > 0) {
+
+                    MessageReceivedEvent event = (MessageReceivedEvent) data.get();
+                    event.getMessage().delete().queue();
+
+                    StringBuilder message = new StringBuilder();
+                    for (int i = 0; i < data.getArgs().length; i++) {
+                        String word = data.getArgs()[i];
+                        message.append(word);
+                        if (i != data.getArgs().length - 1)
+                            message.append(" ");
+
+                    }
+
+                    String title = null;
+                    String description = null;
+                    if (message.toString().startsWith("\"")) {
+                        String temp = message.toString().substring(1);
+                        int index = temp.indexOf("\"");
+                        title = temp.substring(0, index);
+                        description = message.toString().substring(index + 2);
+                    }
+
+
+                    if (title != null) {
+                        embedBuilder.setTitle(title);
+                        embedBuilder.setDescription(description);
+                    } else {
+                        embedBuilder.setDescription(message.toString());
+                    }
+
+                } else {
+                    embedBuilder.setDescription(getPrefix() + "em (description)\n" + getPrefix() + "em \"(title)\" (description)");
+                }
                 return new CoreCommandResponse(data).setEmbed(embedBuilder.build());
             } else {
                 return new CoreCommandResponse(data).setDescription("This is a discord only command.");
